+++ src/CombinationGenerationCounter/entities/CombinationGenerationCounter.entity.go
@@ -1,15 +1,6 @@
 package CombinationGenerationCounterEntity
 
-import "errors"
-
 type CombinationGenerationCounter struct {
-	NumberOfEntries int16  `json:"NumberOfEntries"`
-	Count           *int32 `json:"Count"`
-}
-
-func Validate(combination *CombinationGenerationCounter) error {
-	if combination.NumberOfEntries < 1 || combination.NumberOfEntries > 20 {
-		return errors.New("NumberOfEntries is more than 20 or less than 1")
-	}
-	return nil
+	NumberOfInputs int16 `json:"NumberOfInputs"`
+	Count          int32 `json:"Count"`
 }